FEAT: Add to cart functionality implemented
Implemented add to cart functionality with context. First time the item will be added in the cart, if the item already in the cart then it will just increase the quantity.

Conditionally set the position of navbar to fixed. when user scroll down then it will be on the top all time.